rating-comment fix

+++ src/main/java/com/bookstore/GeekText/model/RatingComment.java
@@ -50,7 +50,7 @@ public class RatingComment{
         this.isbn = isbn;
     }
 
-    public float getRating() {
+    public int getRating() {
         return rating;
     }
